null_blk: add zone support
Add support for exposing a null_blk device through the zone device
model.

Add two new parameters, zoned and zone_size. If zoned is set,
null_blk will register the device as a zoned block device, and
expose a set of zones that must be written sequentially.

The zone_size defines the size of each zone. It defaults to 256MB, as
this follows what read ZBC devices traditionally exposes.

Signed-off-by: Matias Bjørling <matias.bjorling@wdc.com>
Signed-off-by: Bart Van Assche <bart.vanassche@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>